🎨 스크롤 상태를 전역으로 관리
